Solution refactoring
