+++ GameEngine/GameEngine/Engine/Culling/Ouptut/RenderLight.hh
@@ -2,8 +2,8 @@
 
 #include <vector>
 #include <Render/Textures/Texture2D.hh>
-#include <SpacePartitioning/Cullable/Object/PointLight.hh>
-#include <SpacePartitioning/Ouptut/RenderPainter.hh>
+#include <Culling/Cullable/Object/PointLight.hh>
+#include <Culling/Ouptut/RenderPainter.hh>
 #include <glm/glm.hpp>
 
 namespace AGE
