+++ GameEngine/GameEngine/Engine/Culling/PreparableObject.cpp
@@ -1,4 +1,4 @@
-#include <SpacePartitioning/PreparableObject.hh>
+#include <Culling/PreparableObject.hh>
 
 namespace AGE
 {
