+++ GameEngine/GameEngine/Engine/Render/Pipelining/Pipelines/CustomPipeline/DeferredShading.hh
@@ -3,7 +3,7 @@
 #include <Utils/Debug.hpp>
 #include <Utils/OpenGL.hh>
 #include <Render/Pipelining/Pipelines/ARenderingPipeline.hh>
-#include <SpacePartitioning/Ouptut/RenderLight.hh>
+#include <Culling/Ouptut/RenderLight.hh>
 #include <glm/glm.hpp>
 #include <Render/Textures/Texture2D.hh>
