+++ GameEngine/GameEngine/Engine/Threads/RenderScene.cpp
@@ -9,13 +9,13 @@
 #include <glm/gtx/quaternion.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <chrono>
-#include <SpacePartitioning/Cullable/Object/Camera.hh>
-#include <SpacePartitioning/Cullable/Object/PointLight.hh>
-#include <SpacePartitioning/Cullable/Object/Mesh.hh>
-#include <SpacePartitioning/Ouptut/RenderCamera.hh>
-#include <SpacePartitioning/Ouptut/RenderLight.hh>
-#include <SpacePartitioning/Ouptut/RenderPainter.hh>
-#include <SpacePartitioning/Ouptut/RenderPipeline.hh>
+#include <Culling/Cullable/Object/Camera.hh>
+#include <Culling/Cullable/Object/PointLight.hh>
+#include <Culling/Cullable/Object/Mesh.hh>
+#include <Culling/Ouptut/RenderCamera.hh>
+#include <Culling/Ouptut/RenderLight.hh>
+#include <Culling/Ouptut/RenderPainter.hh>
+#include <Culling/Ouptut/RenderPipeline.hh>
 #include <Threads/RenderThread.hpp>
 #include <Utils/Debug.hpp>
 #include <Core/Link.hpp>
